requirement.txt
discord.py>=2.3.0
yt-dlp>=2023.11.16
bilibili-api-python>=15.0.0
openai>=1.0.0
python-dotenv>=1.0.0
psutil>=5.9.0